Chore: removes todos
