+++ packages/grafana-e2e-selectors/package.json
@@ -18,6 +18,7 @@
   "main": "src/index.ts",
   "scripts": {
     "build": "grafana-toolkit package:build --scope=e2e-selectors",
+    "bundle": "rollup -c rollup.config.ts",
     "clean": "rimraf ./dist ./compiled",
     "docsExtract": "mkdir -p ../../reports/docs && api-extractor run 2>&1 | tee ../../reports/docs/$(basename $(pwd)).log",
     "lint": "eslint src/ --ext=.js,.ts,.tsx",
@@ -25,6 +26,16 @@
   },
   "devDependencies": {
     "@types/node": "13.7.7",
+    "@rollup/plugin-commonjs": "11.0.2",
+    "@rollup/plugin-node-resolve": "7.1.1",
+    "@types/rollup-plugin-visualizer": "2.6.0",
+    "@types/systemjs": "^0.20.6",
+    "pretty-format": "25.1.0",
+    "rollup": "2.0.6",
+    "rollup-plugin-sourcemaps": "0.5.0",
+    "rollup-plugin-terser": "5.3.0",
+    "rollup-plugin-typescript2": "0.26.0",
+    "rollup-plugin-visualizer": "3.3.1",
     "ts-loader": "6.2.1",
     "ts-node": "8.8.1"
   },
